+++ src/components/Tag/Tag.tsx
@@ -1,22 +1,22 @@
-import { CharacterType } from '@/types/id-cards.type';
+import { CharacterNameType, CharacterType } from '@/types/idCard';
 
 type TagProps = {
-  type: CharacterType;
+  type: CharacterNameType;
   label: string;
 };
 
-const colors: Record<CharacterType, string> = {
+const colors: Record<CharacterNameType, string> = {
   BUDDY: 'text-buddy-700 bg-buddy-200 border-buddy-400',
   TOBBY: 'text-tobby-700 bg-tobby-200 border-tobby-400',
   PIPI: 'text-pipi-700 bg-pipi-200 border-pipi-400',
   TRUE: 'text-true-700 bg-true-200 border-true-400',
 };
 
-const getCharacterColor = (type: CharacterType) => {
+const getCharacterColor = (type: CharacterNameType) => {
   return `${colors[type]}`;
 };
 
-const getClassName = (type: CharacterType) => {
+const getClassName = (type: CharacterNameType) => {
   return `${getCharacterColor(
     type,
   )} inline-block rounded border border-solid px-2 py-1 text-detail font-medium`;
