+++ src/lib/api/config/requestUrl.ts
@@ -2,7 +2,7 @@ export const ROOT_API_URL = process.env.ROOT_API_URL;
 
 export const REQUEST_API = {
   TESTER: '/tester',
-  IDCARDS: {
+  IDCARD: {
     DETAIL: '/id-cards',
   },
 };
