+++ src/mocks/idCard/idCard.mock.ts
@@ -1,9 +1,9 @@
 import { faker } from '@faker-js/faker/locale/ko';
 
-import { IdCardType } from '@/types/id-cards.type';
+import { IdCardDetailType } from '@/types/idCard';
 
-export const createIdCard = (): IdCardType => ({
-  idCardId: faker.string.uuid(),
+export const createIdCard = (): IdCardDetailType => ({
+  idCardId: faker.number.int(),
   nickname: faker.person.fullName(),
   profileImageUrl: faker.image.avatar(),
   aboutMe: faker.lorem.paragraph(),
