+++ src/mocks/idCard/idCard.mockHandler.ts
@@ -1,13 +1,13 @@
 import { rest } from 'msw';
 
 import { REQUEST_API, ROOT_API_URL } from '@/lib/api/config/requestUrl';
-import { createIdCard } from '@/mocks/id-cards/id-cards.mock';
+import { createIdCard } from '@/mocks/idCard/idCard.mock';
 
-const idCardsMockHandler = [
+const idCardMockHandler = [
   // GET: api/id-cards/{idCardsId}
-  rest.get(`${ROOT_API_URL}${REQUEST_API.IDCARDS.DETAIL}/:id`, (req, res, ctx) => {
+  rest.get(`${ROOT_API_URL}${REQUEST_API.IDCARD.DETAIL}/:id`, (req, res, ctx) => {
     return res(ctx.status(200), ctx.json({ idCardDetailsDto: createIdCard() }));
   }),
 ];
 
-export default idCardsMockHandler;
+export default idCardMockHandler;
